// apps/web/src/api/models/mesaBindingCreate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * MESA Law API
 * OpenAPI spec version: 0.1.0
 */

export interface MesaBindingCreate {
  /**
     * @minLength 1
     * @maxLength 128
     */
  agent_id: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  dataset_id: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  mesa_tenant_id: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  workspace_id: string;
}